DALL E - image gen/requirements.txt
numpy==1.24.1
openai==0.26.0
Pillow==9.4.0
streamlit==1.16.0
urllib3==1.26.13
requests==2.28.1
